back/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
axios==0.4.0
blinker==1.6.3
Brotli==1.1.0
cachetools==5.5.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
environ==1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==13.3.4
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.28.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.0.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huepy==1.2.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
livereload==2.7.0
locust==2.17.0
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.0.7
numpy==2.0.2
oauthlib==3.2.2
openai==1.54.4
openpyxl==3.1.5
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.6
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.10.1
pyparsing==3.2.0
pyphen==0.17.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.1
redis==5.2.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.4
roundrobin==0.0.4
rpds-py==0.21.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
textstat==0.7.4
toposort==1.7
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vite==1.5.2
vue==0.0.1
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
